ensure complete tiles when using the scene footprints as geometry
Before incomplete tiles could be created at the ends of a data take group because all tiles overlapping with the scenes were considered.
Now a second scene search is performed with the preselected tiles as input geometry and an extended time range so that pre- and succeeding scenes are selected as well.